renaming files and adding types file
