+++ src/view/export-block/types.ts
@@ -1,0 +1,3 @@
+export type ExportBlockProps = {
+  exportCode: string | Array<any>
+}
