+++ src/view/image-info-panel/index.css
@@ -19,13 +19,3 @@
 .zoom {
   padding: 0 12px;
 }
-
-/* .image-size:after {
-  content: '';
-  position: absolute;
-  right: -8px;
-  top: 0;
-  width: 2px;
-  height: 40px;
-  background-color: #282C34;
-} */
