+++ src/view/image-info-panel/types.ts
@@ -1,0 +1,4 @@
+export type ImageInfoPanelProps = {
+  zoom: number,
+  imageSize: string,
+}
