+++ src/view/image-zone/types.ts
@@ -1,0 +1,9 @@
+export type ImageZoneProps = {
+  imageSrc: string,
+}
+
+export type scaleOptions = {
+  stageScale?: number,
+  stageX?: number,
+  stageY?: number
+}
