+++ src/view/intro-zone/index.css
