+++ src/view/intro-zone/index.tsx
