+++ src/view/matter-demo/types.ts
@@ -1,0 +1,5 @@
+import { Vector } from "matter-js"
+
+export type MatterDemoProps = {
+  vertices: Vector[]
+}
