+++ src/view/shape-controls/types.ts
@@ -1,0 +1,6 @@
+import { World } from "matter-js";
+
+export type ShapeControlsProps = {
+  world: World,
+  isShapeConvex?: boolean
+}
