+++ src/view/tool-icon/types.ts
@@ -1,0 +1,5 @@
+import { EditingTools } from "../../types";
+
+export type ToolIconProps = {
+  alias: EditingTools,
+}
