+++ src/view/ui-layer/index.tsx
@@ -1,15 +1,15 @@
-import { ToolPanel } from '../tool-panel/tool-panel'
+import { ToolPanel } from '../tool-panel'
 import { $lineCoords, $loadedFile, $shapeSettings, $zoomValue } from '../../model/store'
 import { useStore } from 'effector-react'
-import { UtilityBtn } from '../utility-btn/utility-btn'
+import { UtilityBtn } from '../utility-btn'
 import { Modal, Snackbar } from '@mui/material'
 import { useState } from 'react'
-import { ExportBlock } from '../export-block/export-block'
+import { ExportBlock } from '../export-block/'
 import './ui-layer.css'
 import { setFileLoaded, showEmulation } from '../../model/events'
 import { SNACKBAR_MESSAGE } from '../../data'
 import { createExportCode } from '../../utils/create-export-code'
-import { ImageInfoPanel } from '../image-info-panel/image-info-panel'
+import { ImageInfoPanel } from '../image-info-panel'
 import useImage from 'use-image'
 
 export const UILayer = (): JSX.Element => {
