+++ src/view/utility-btn/types.ts
@@ -1,0 +1,4 @@
+export type UtilityBtnProps = {
+  onClick?: () => void,
+  alias: string,
+}
